+++ lib/components/UploadFiles/index.ts
@@ -1,2 +1,2 @@
-export type { UploadFilesProps } from "./types";
+export type { IUploadFilesProps } from "./types";
 export { UploadFiles } from "./UploadFiles.tsx";