+++ lib/test-stpr-ui-kit.ts
@@ -1,49 +1,58 @@
 // Включение include-tokens в бандл (для объявления переменных)
 import "../src/styles/include-tokens.scss";
 
-import { Accordion } from "./components/Accordion";
-import { ApplyButtons } from "./components/ApplyButtons";
-import { AuthProtected } from "./components/AuthProtected";
-import { BaseTooltip } from "./components/BaseTooltip";
-import { Breadcrumb, type TCrumbItem } from "./components/Breadcrumb";
-import { Button } from "./components/Button";
-import type { TOnChangeCheckbox } from "./components/Checkbox";
-import { Checkbox } from "./components/Checkbox";
-import { Confirm } from "./components/Confirm";
-import type { TContextMenuOption } from "./components/ContextMenu";
-import { ContextMenu } from "./components/ContextMenu";
-import type { TOnChangeDatePicker } from "./components/DatePicker";
-import { DatePicker } from "./components/DatePicker";
-import type { TOnChangeDatePickerInput } from "./components/DatePickerInput";
-import { DatePickerInput } from "./components/DatePickerInput";
-import { DefaultDropzone } from "./components/DefaultDropzone";
-import { Dropdown } from "./components/Dropdown";
-import { EllipsisTextWithTooltip } from "./components/EllipsisTextWithTooltip";
-import { Form } from "./components/Form";
-import { EIconName, Icon } from "./components/Icons";
-import type { TOnBlurInput, TOnChangeInput } from "./components/Input";
-import { Input } from "./components/Input";
-import { Label } from "./components/Label";
-import { Layer } from "./components/Layer";
-import { MediaContext, MediaContextProvider } from "./components/MediaContextProvider";
-import { Modal, useModal } from "./components/Modal";
-import { Portal } from "./components/Portal";
-import { ProgressWrapper, useAnimatedValue } from "./components/ProgressWrapper";
-import type { TOnBlurSelect, TOnChangeSelect, TSelectOption } from "./components/Select";
-import { Select } from "./components/Select";
-import { Skeleton } from "./components/Skeleton";
-import { Spinner } from "./components/Spinner";
-import type { TableProps, TClickOnCellAction, TColumn } from "./components/Table";
-import { Table, useTableActions } from "./components/Table";
-import type { TabsProps, TPaneItem } from "./components/Tabs";
-import { Tabs } from "./components/Tabs";
-import { Text } from "./components/Text";
-import type { TOnChangeTextarea } from "./components/Textarea";
-import { Textarea } from "./components/Textarea";
-import { TextWithLabel } from "./components/TextWithLabel";
-import { ETooltipPosition, InfoTooltip, Tooltip } from "./components/Tooltip";
-import { UploadFiles } from "./components/UploadFiles";
-import { ViewImageModal } from "./components/ViewImageModal";
+import { Accordion } from "@components/Accordion";
+import { ApplyButtons } from "@components/ApplyButtons";
+import { AuthProtected } from "@components/AuthProtected";
+import { BaseTooltip } from "@components/BaseTooltip";
+import { Breadcrumb, type TCrumbItem } from "@components/Breadcrumb";
+import { Button } from "@components/Button";
+import type { TOnChangeCheckbox } from "@components/Checkbox";
+import { Checkbox } from "@components/Checkbox";
+import { Confirm } from "@components/Confirm";
+import type { TContextMenuOption } from "@components/ContextMenu";
+import { ContextMenu } from "@components/ContextMenu";
+import type { TOnChangeDatePicker } from "@components/DatePicker";
+import { DatePicker } from "@components/DatePicker";
+import type { TOnChangeDatePickerInput } from "@components/DatePickerInput";
+import { DatePickerInput } from "@components/DatePickerInput";
+import { DefaultDropzone } from "@components/DefaultDropzone";
+import type { TDropdownListItem } from "@components/Dropdown";
+import { Dropdown } from "@components/Dropdown";
+import { EllipsisTextWithTooltip } from "@components/EllipsisTextWithTooltip";
+import { Form } from "@components/Form";
+import { EIconName, Icon } from "@components/Icons";
+import type { IIfcPreviewProps } from "@components/IfcPreview";
+import { IfcPreview } from "@components/IfcPreview";
+import type { IInputProps, TOnBlurInput, TOnChangeInput } from "@components/Input";
+import { Input } from "@components/Input";
+import { Label } from "@components/Label";
+import { Layer } from "@components/Layer";
+import { MediaContext, MediaContextProvider } from "@components/MediaContextProvider";
+import { Modal, useModal } from "@components/Modal";
+import { Portal } from "@components/Portal";
+import { ProgressWrapper, useAnimatedValue } from "@components/ProgressWrapper";
+import type {
+  ISelectProps,
+  TOnBlurSelect,
+  TOnChangeSelect,
+  TSelectOption,
+} from "@components/Select";
+import { Select } from "@components/Select";
+import { Skeleton } from "@components/Skeleton";
+import { Spinner } from "@components/Spinner";
+import type { TableProps, TClickOnCellAction, TColumn } from "@components/Table";
+import { Table, useTableActions } from "@components/Table";
+import type { TabsProps, TPaneItem } from "@components/Tabs";
+import { Tabs } from "@components/Tabs";
+import { Text } from "@components/Text";
+import type { ITextareaProps, TOnChangeTextarea } from "@components/Textarea";
+import { Textarea } from "@components/Textarea";
+import { TextWithLabel } from "@components/TextWithLabel";
+import { ETooltipPosition, InfoTooltip, Tooltip } from "@components/Tooltip";
+import type { IUploadFilesProps } from "@components/UploadFiles";
+import { UploadFiles } from "@components/UploadFiles";
+import { ViewImageModal } from "@components/ViewImageModal";
 
 export {
   Accordion,
@@ -64,8 +73,14 @@ export {
   ETooltipPosition,
   Form,
   Icon,
+  IfcPreview,
+  IIfcPreviewProps,
+  IInputProps,
   InfoTooltip,
   Input,
+  ISelectProps,
+  ITextareaProps,
+  IUploadFilesProps,
   Label,
   Layer,
   MediaContext,
@@ -84,6 +99,7 @@ export {
   TColumn,
   TContextMenuOption,
   TCrumbItem,
+  TDropdownListItem,
   Text,
   Textarea,
   TextWithLabel,